Split ordering functionality into new file from smart playlists
